# apps/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "0x1audio",
    "summary": "Private music catalog with search, chat, and playlists.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/search/text": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search Text",
        "operationId": "searchText",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchTextBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/audio": {
      "post": {
        "tags": [
          "search"
        ],
        "summary": "Search Audio",
        "operationId": "searchAudio",
        "parameters": [
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 6,
              "title": "Top K"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_searchAudio"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/similar/{track_id}": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Similar Tracks",
        "operationId": "similarTracks",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Track Id"
            }
          },
          {
            "name": "top_k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 6,
              "title": "Top K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistRefOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ArtistRefOut"
      },
...
      "Body_searchAudio": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_searchAudio"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchHitOut": {
        "properties": {
          "track": {
            "$ref": "#/components/schemas/TrackOut"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "track_score": {
            "type": "number",
            "title": "Track Score"
          },
          "best_segment_score": {
            "type": "number",
            "title": "Best Segment Score"
          },
          "segment_coverage": {
            "type": "number",
            "title": "Segment Coverage"
          },
          "match_scope": {
            "type": "string",
            "title": "Match Scope"
          },
          "matched_segment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Segment Ids"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          }
        },
        "type": "object",
        "required": [
          "track",
          "score",
          "track_score",
          "best_segment_score",
          "segment_coverage",
          "match_scope",
          "matched_segment_ids",
          "reasons"
        ],
        "title": "SearchHitOut"
      },
      "SearchMode": {
        "type": "string",
        "enum": [
          "tracks",
          "segments"
        ],
        "title": "SearchMode"
      },
      "SearchResponseOut": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchHitOut"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "SearchResponseOut"
      },
      "SearchTextBody": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query"
          },
          "negative_query": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Negative Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 6
          },
          "mode": {
            "$ref": "#/components/schemas/SearchMode",
            "default": "tracks"
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchTextBody"
      },
...
      "TrackOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/TrackStatus"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "imported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Imported At"
          },
          "indexed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexed At"
          },
          "cover_color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Color"
          },
          "has_cover": {
            "type": "boolean",
            "title": "Has Cover",
            "default": false
          },
          "duration_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration S"
          },
          "is_instrumental": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Instrumental"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistRefOut"
            },
            "type": "array",
            "title": "Artists",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "artist",
          "original_filename",
          "content_type",
          "size_bytes",
          "status",
          "error_message",
          "imported_at",
          "indexed_at"
        ],
        "title": "TrackOut"
      },
      "TrackStatus": {
        "type": "string",
        "enum": [
          "uploading",
          "queued",
          "indexing",
          "ready",
          "failed"
        ],
        "title": "TrackStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}